Crate mdbook_linkcheck

A mdbook backend which will check all links in a document are valid.

The link-checking process has roughly three stages:

Find all the links in a body of markdown text (see [extract_links])

Validate all the links we've found, taking into account cached results and configuration options

Cache the results in the output directory for reuse by step 2 in the next round

Emit errors/warnings to the user

Structs

Cache
A cache used to avoid unnecessary web requests.

Config
The configuration options available with this backend.

IncompleteLink
A potential link that has a broken reference (e.g [foo] when there is no [foo]: ... entry at the bottom).

InvalidLink
An invalid [Link] and the [Reason] for why it isn't valid.

Link
A single link, and where it was found in the parent document.

UnknownScheme
An unknown [Uri::scheme_str()].

ValidationOutcome
The outcome of validating a set of links.

Enums

Reason
Why is this [Link] invalid?

Constants

COMPATIBLE_MDBOOK_VERSIONS
A semver range specifying which versions of mdbook this crate supports.

Functions

extract_links
Search every file in the [Files] and collate all the links that are found.

get_config
Get the configuration used by mdbook-linkcheck.

load_files_into_memory
A helper for reading the chapters of a [Book] into memory.

validate
Try to validate the provided [Link]s.

version_check
Check whether this library is compatible with the provided version string.