Trait mdbook_linkcheck::errors::BrokenLink

pub trait BrokenLink: Fail {
    fn chapter(&self) -> &Path;

    fn line(&self) -> usize;

    fn as_fail(&self) -> &dyn Fail;
}

Some arbitrary broken link which occurs at a specific line in a chapter. The Display impl should state why the link is "broken".

Required methods

fn chapter(&self) -> &Path

Which chapter it was in.

fn line(&self) -> usize

The line this error occurred on.

fn as_fail(&self) -> &dyn Fail

Loading content...